+++ src/main/kotlin/dev/aasmart/game/ConnectFourGame.kt
@@ -2,8 +2,7 @@ package dev.aasmart.game
 
 import dev.aasmart.dao.games.GamesFacade
 import dev.aasmart.models.*
-import dev.aasmart.models.games.GameState
-import dev.aasmart.models.games.GameTile
+import dev.aasmart.models.games.*
 import dev.aasmart.utils.*
 import io.ktor.websocket.*
 import kotlinx.serialization.encodeToString
